\name{print.cquad}

\alias{print.cquad}

\title{Print output for class cquad}

\description{Print output for class cquad and output provided by cquad_basic, cquad_equ, cquad_ext, cquad_pseudo}

\usage{
	\method{print}{cquad}(x, ...)
}

\arguments{
	\item{x}{output of class cquad}
	\item{...}{further arguments passed to or from other methods}
}

\author{
  Francesco Bartolucci (University of Perugia), Claudia Pigini (University of Ancona "Politecnica delle Marche")
}