\name{summary.cquad}

\alias{summary.cquad}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Summary for class cquad}

\description{Summarize the output for class cquad provided by cquad_basic, cquad_equ, cquad_ext, cquad_pseudo}

\usage{
	\method{summary}{cquad}(object, ...)
}

\arguments{
	\item{object}{output of class cquad}
	\item{...}{further arguments passed to or from other methods}
}

\author{
	Francesco Bartolucci (University of Perugia), Claudia Pigini (University of Ancona "Politecnica delle Marche")
}